requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
Django==4.2.1
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap-form==3.4
django-bootstrap4==23.1
django-crispy-forms==2.0
django-filter==23.2
django-globals==0.3.2
django-html==0.1.0
django-htmx==1.14.0
django-tables2==2.5.3
django-widget-tweaks==1.4.12
mysqlclient==2.1.1
pydantic==1.10.7
pyserial==3.5
serial-weighing-scale==0.0.6
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
